Cleaned up Javadoc and README.md.
I experimented with using pandoc to generate Javadoc HTML from
the README.md markdown. That didn't work too well, because the
HTML generated by pandoc is very hard to read. Javadoc comments
should be readable as source code as well as when translated
to HTML. So for now I am maintaining README.md and the Gridder
doc comment separately, and trying to keep them reasonably
consistent.